+++ NEWS.md
@@ -1,3 +1,9 @@
 # NetworkExtinction 0.1.1
 
 * Added a `NEWS.md` file to track changes to the package.
+
+# NetworkExtinction 0.1.2
+
+* Fixed parameters in `degree_distribution`, the truncated parameters were not present
+* Fixed parameters in `degree_distribution`, the intercept for power law was not present
+* Fixed `degree_distribution` to select one model per family